+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 * Update AWS Resource specification to v3.3.0
+* Utilities to create inline AWS Lambda functions
 
 ## [v0.6.3] - 2019-03-14
 ### Changed